rename and make config folder
