+++ Src/ExternalModelsAPI/Message.py
@@ -1,14 +1,13 @@
 # you have many data objects, but only 
 # the following 
-from abc import ABC
-from typing import un
-class TextContent:
-    def __init__(self, text):
-        self.type = "text"
-        self.text = text
 
-    def to_dict(self):
-        return {"type": self.type, "text": self.text}
+# class TextContent:
+#     def __init__(self, text):
+#         self.type = "text"
+#         self.text = text
+
+#     def to_dict(self):
+#         return {"type": self.type, "text": self.text}
 
 class Message:
     """
