+++ Src/ExternalModelsAPI/unit_test.py
@@ -1,6 +1,6 @@
 import unittest
-from messages import Message
-from prompt_list import Prompt_List
+from Message import Message
+from Prompt_List import Prompt_List
 #testing
 class TestMessage(unittest.TestCase):
     def test_proper_initialization(self):
